#91: Create new templates for Github Workflows - prepare testcontainer
