+++ .gitattributes
@@ -1,2 +1,3 @@
 poetry.lock linguist-generated=true
 exasol_script_languages_container_ci_setup/version.py linguist-generated=true
+.github/workflows/slc_ci*.yml linguist-generated=true
