+++ .github/workflows/slc_ci.yml
@@ -1,23 +1,20 @@
+# ---- AUTOMATICALLY GENERATED FILE ----
+# ---- DO NOT EDIT MANUALLY, BUT USE PYTHON MODULE "exasol.slc_ci_setup" ----
+# ---- from https://github.com/exasol/script-languages-container-ci-setup TO UPDATE ----
+
+
 name: SLC-CI
 
 on:
   push:
     branches:
       - master
-      - develop
       - main
+      - develop
+      - 'rebuild/**'
   pull_request:
 
 jobs:
-  dummy:
-    runs-on: ubuntu-22.04
-    steps:
-      - uses: actions/checkout@v4
-      - name: Setup Python & Poetry Environment
-        uses: exasol/python-toolbox/.github/actions/python-environment@1.0.1
-        with:
-          python-version: '3.10'
-          poetry-version: '2.1.2'
-
-      - name: Dummy
-        run: echo Test
+  prepare-test-container:
+    uses: ./.github/workflows/slc_ci_prepare_test_container.yml
+    secrets: inherit
