+++ .github/workflows/slc_ci_self_check.yml
@@ -1,3 +1,8 @@
+# ---- AUTOMATICALLY GENERATED FILE ----
+# ---- DO NOT EDIT MANUALLY, BUT USE PYTHON MODULE "exasol.slc_ci_setup" ----
+# ---- from https://github.com/exasol/script-languages-container-ci-setup TO UPDATE ----
+
+
 name: SLC-CI-Self-Check
 
 on:
