--- exasol/slc_ci_setup/lib/render_template.py
@@ -1,11 +1,0 @@
-import jinja2
-
-
-def render_template(template: str, **kwargs):
-    env = jinja2.Environment(
-        loader=jinja2.PackageLoader("exasol.slc_ci_setup"),
-        autoescape=jinja2.select_autoescape(),
-        keep_trailing_newline=True,
-    )
-    t = env.get_template(template)
-    return t.render(**kwargs)
