+++ exasol/slc_ci_setup/templates/slc_ci_prepare_test_container.yml.tmpl
@@ -1,0 +1,29 @@
+{% include "heading.tmpl" %}
+{% raw %}
+name: SLC-CI-Prepare-Test-Container
+
+on:
+  workflow_call:
+    secrets:
+      DOCKER_USERNAME:
+        required: true
+      DOCKER_PASSWORD:
+        required: true
+jobs:
+  prepare-test-container:
+    runs-on: ubuntu-24.04
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup Python & Poetry Environment
+        uses: exasol/python-toolbox/.github/actions/python-environment@1.2.0
+        with:
+          python-version: '3.10'
+          poetry-version: '2.1.2'
+      - name: prepare test container
+        run: |
+            echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin
+            poetry run -- exaslc-ci prepare-test-container --commit-sha "$COMMIT_SHA" --docker-user "$DOCKER_USERNAME" --docker-password "$DOCKER_PASSWORD"
+        env:
+          DOCKER_USERNAME: ${{ secrets.DOCKER_USERNAME }}
+          DOCKER_PASSWORD: ${{ secrets.DOCKER_PASSWORD }}
+          COMMIT_SHA: ${{ github.sha }}{% endraw %}
