+++ exasol/slc_ci_setup/templates/slc_ci_self_check.yml.tmpl
@@ -1,3 +1,5 @@
+{% include "heading.tmpl" %}
+{% raw %}
 name: SLC-CI-Self-Check
 
 on:
@@ -29,4 +31,4 @@ jobs:
         run: |
           git status --porcelain=v1 -uno
           git diff --cached; git diff --cached --summary;
-          [ -z "$(git status --porcelain=v1 -uno 2>/dev/null)" ]
+          [ -z "$(git status --porcelain=v1 -uno 2>/dev/null)" ]{% endraw %}
