+++ flavors/integration_test/ci.json
@@ -1,12 +1,7 @@
 {
   "build_runner": "ubuntu-22.04",
   "test_config": {
-    "test_runner": "ubuntu-22.04",
-    "test_sets": [
-      {
-        "name": "integration-test",
-        "folders": ["some_tests"]
-      }
-    ]
+    "default_test_runner": "ubuntu-22.04",
+    "test_sets": []
   }
 }
